Clean up MIST dataset conversion entry point.
